Chore: Restructuring files
