+++ src/Pages/AppraisalActivity/ActivityForm.js
@@ -1,71 +1,24 @@
 // src/components/AddActivityForm.js
 import { Formik, Form } from 'formik';
-import FormikControl from './formikControl';
+import FormikControl from './FormikControl';
 import validationSchema from './Schema';
 import Configs from '../../CommonConfigs/Configs';
 
 function AddActivityForm () {
-    const initialValues = {
-      period: '',
-      perspective: '',
-      ssMartaObjectives: '',
-      initiative: '',
-      measurableActivities: '',
-      implementations: '',
-      date: '',
-      comments: '',
-      stakeholders: '',
-      evidence: ''
-    };
-   
-    const perspectiveDropdownOptions = [
-      { key: 'Select a perspective', value: '' },
-      { key: 'Organisation', value: 'organisation' },
-      { key: 'Business', value: 'business' },
-      { key: 'Financial', value: 'financial' },
-      { key: 'Client', value: 'client' }
-    ];
-    
-    const ssmartaObjectiveDropdownOptions = [
-      { key: 'Select an objective', value: '' },
-      { key: 'SO1', value: 'so1' },
-      { key: 'SO2', value: 'so2' },
-      { key: 'SO3', value: 'so3' },
-      { key: 'SO4', value: 'so4' }
-    ];
-    
-    const initiativeDropdownOptions = [
-      { key: 'Select an initiative', value: '' },
-      { key: 'I1', value: 'i1' },
-      { key: 'I2', value: 'i2' },
-      { key: 'I3', value: 'i3' },
-      { key: 'I4', value: 'i4' }
-    ];
-    
-    const measurableActivityDropdownOptions = [
-      { key: 'Select a measurable activity', value: '' },
-      { key: 'MA1', value: 'ma1' },
-      { key: 'MA2', value: 'ma2' },
-      { key: 'MA3', value: 'ma3' },
-      { key: 'MA4', value: 'ma4' }
-    ];
-    
-    const onSubmit = (values, onSubmitProps)  => {
+      const onSubmit = (values, onSubmitProps)  => {
         const formData = {
           ...values,
           evidence: values.evidence.split(',').map(item => item.trim()), // Assuming evidence is a comma-separated string
           date: new Date(values.date).toISOString() // Ensure date is a string
         }
-        console.log('Form Data:', formData) // Debug the form data
         onSubmit(formData)
         onSubmitProps.setSubmitting(false)
         onSubmitProps.resetForm()
-      
     } 
     
   return (
     <Formik
-      initialValues={initialValues}
+      initialValues={Configs.initialValues}
       validationSchema={validationSchema}
       onSubmit={(values, onSubmitProps) => {
         const formData = {
@@ -101,7 +54,7 @@ function AddActivityForm () {
             control='select'
             label='Perspective'
             name='perspective'
-            options={perspectiveDropdownOptions}
+            options={Configs.perspectiveDropdownOptions}
           />
           </div>
 
@@ -110,7 +63,7 @@ function AddActivityForm () {
             control='select'
             label='Ssmarta Objectives'
             name='ssMartaObjectives'
-            options={ssmartaObjectiveDropdownOptions}
+            options={Configs.ssmartaObjectiveDropdownOptions}
           />
           </div>
 
@@ -119,7 +72,7 @@ function AddActivityForm () {
             control='select'
             label='Initiative'
             name='initiative'
-            options={initiativeDropdownOptions}
+            options={Configs.initiativeDropdownOptions}
           />
           </div>
 
@@ -128,7 +81,7 @@ function AddActivityForm () {
             control='select'
             label='Measurable Activities'
             name='measurableActivities'
-            options={measurableActivityDropdownOptions}
+            options={Configs.measurableActivityDropdownOptions}
           />
           </div>
